feat: migrate main landing page to web_new
- Migrated index.html content to src/routes/+page.svelte
- Moved registration form to src/routes/registrace/+page.svelte
- Created Navbar and Footer components in src/lib/components
- Updated src/routes/+layout.svelte with global styles and navigation
- Copied assets (logos, propozice) to static/
- Updated event handlers to Svelte 5 syntax (onclick)
- Fixed accessibility and TypeScript issues in migrated components
