+++ dev_server.log
@@ -1,0 +1,30 @@
+
+> web_new@0.0.1 dev
+> vite dev
+
+8:51:53 AM [vite] (client) Forced re-optimization of dependencies
+
+  VITE v8.0.8  ready in 1230 ms
+
+  ➜  Local:   http://localhost:5173/
+  ➜  Network: use --host to expose
+8:54:11 AM [vite-plugin-svelte] src/routes/registrace/+page.svelte:307:43 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
+https://svelte.dev/e/event_directive_deprecated
+8:54:11 AM [vite-plugin-svelte] src/routes/registrace/+page.svelte:317:6 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
+https://svelte.dev/e/event_directive_deprecated
+8:54:11 AM [vite-plugin-svelte] src/routes/registrace/+page.svelte:629:66 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
+https://svelte.dev/e/event_directive_deprecated
+8:54:11 AM [vite-plugin-svelte] src/routes/registrace/+page.svelte:634:73 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
+https://svelte.dev/e/event_directive_deprecated
+8:54:11 AM [vite-plugin-svelte] src/routes/registrace/+page.svelte:638:73 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
+https://svelte.dev/e/event_directive_deprecated
+8:54:12 AM [vite-plugin-svelte] src/routes/registrace/+page.svelte:307:43 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
+https://svelte.dev/e/event_directive_deprecated
+8:54:12 AM [vite-plugin-svelte] src/routes/registrace/+page.svelte:317:6 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
+https://svelte.dev/e/event_directive_deprecated
+8:54:12 AM [vite-plugin-svelte] src/routes/registrace/+page.svelte:629:66 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
+https://svelte.dev/e/event_directive_deprecated
+8:54:12 AM [vite-plugin-svelte] src/routes/registrace/+page.svelte:634:73 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
+https://svelte.dev/e/event_directive_deprecated
+8:54:12 AM [vite-plugin-svelte] src/routes/registrace/+page.svelte:638:73 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
+https://svelte.dev/e/event_directive_deprecated
