--- web_new/dev_server.log
@@ -1,30 +1,0 @@
-
-> web_new@0.0.1 dev
-> vite dev
-
-4:19:12 PM [vite] (client) Forced re-optimization of dependencies
-
-  VITE v8.0.8  ready in 1122 ms
-
-  ➜  Local:   http://localhost:5173/
-  ➜  Network: use --host to expose
-4:19:58 PM [vite-plugin-svelte] src/routes/+page.svelte:299:43 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
-https://svelte.dev/e/event_directive_deprecated
-4:19:58 PM [vite-plugin-svelte] src/routes/+page.svelte:309:6 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
-https://svelte.dev/e/event_directive_deprecated
-4:19:58 PM [vite-plugin-svelte] src/routes/+page.svelte:621:66 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
-https://svelte.dev/e/event_directive_deprecated
-4:19:58 PM [vite-plugin-svelte] src/routes/+page.svelte:626:73 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
-https://svelte.dev/e/event_directive_deprecated
-4:19:58 PM [vite-plugin-svelte] src/routes/+page.svelte:630:73 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
-https://svelte.dev/e/event_directive_deprecated
-4:20:09 PM [vite-plugin-svelte] src/routes/+page.svelte:299:43 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
-https://svelte.dev/e/event_directive_deprecated
-4:20:09 PM [vite-plugin-svelte] src/routes/+page.svelte:309:6 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
-https://svelte.dev/e/event_directive_deprecated
-4:20:09 PM [vite-plugin-svelte] src/routes/+page.svelte:621:66 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
-https://svelte.dev/e/event_directive_deprecated
-4:20:09 PM [vite-plugin-svelte] src/routes/+page.svelte:626:73 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
-https://svelte.dev/e/event_directive_deprecated
-4:20:09 PM [vite-plugin-svelte] src/routes/+page.svelte:630:73 Using `on:click` to listen to the click event is deprecated. Use the event attribute `onclick` instead
-https://svelte.dev/e/event_directive_deprecated
