+++ web_new/src/routes/+layout.svelte
@@ -1,11 +1,38 @@
 <script>
-	import favicon from '$lib/assets/favicon.svg';
+	import Navbar from '$lib/components/Navbar.svelte';
+	import Footer from '$lib/components/Footer.svelte';
+	import '../app.css';
 
 	let { children } = $props();
 </script>
 
 <svelte:head>
-	<link rel="icon" href={favicon} />
+	<link rel="icon" href="/favicon.png" />
+	<link href="https://fonts.googleapis.com/css2?family=Open+Sans:wght@400;700&display=swap" rel="stylesheet">
 </svelte:head>
 
-{@render children()}
+<Navbar />
+
+<main>
+	{@render children()}
+</main>
+
+<Footer />
+
+<style>
+	:global(html) {
+		scroll-behavior: smooth;
+	}
+	:global(body) {
+		margin: 0;
+		background-color: #303b4a;
+		color: #45cece;
+		font-family: 'Open Sans', 'Trebuchet MS', sans-serif;
+		min-height: 100vh;
+		display: flex;
+		flex-direction: column;
+	}
+	main {
+		flex: 1;
+	}
+</style>
